Set settings for jira provider from composition root
