+++ src/Tinkoff.ISA.Providers.Jira/Tinkoff.ISA.Providers.Jira.csproj
@@ -8,7 +8,6 @@
       <PackageReference Include="Atlassian.SDK" Version="10.5.0" />
       <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="2.2.0" />
       <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.2.0" />
-      <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="2.2.0" />
     </ItemGroup>
 
     <ItemGroup>
