+++ src/Tinkoff.ISA.Scheduler/Tinkoff.ISA.Scheduler.csproj
@@ -21,6 +21,8 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="2.2.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.FileExtensions" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.2.0" />
@@ -32,6 +34,7 @@
     <ProjectReference Include="..\Tinkoff.ISA.AppLayer\Tinkoff.ISA.AppLayer.csproj" />
     <ProjectReference Include="..\Tinkoff.ISA.DAL\Tinkoff.ISA.DAL.csproj" />
     <ProjectReference Include="..\Tinkoff.ISA.Infrastructure\Tinkoff.ISA.Infrastructure.csproj" />
+    <ProjectReference Include="..\Tinkoff.ISA.Providers.Jira\Tinkoff.ISA.Providers.Jira.csproj" />
   </ItemGroup>
 
   <ItemGroup>
